+++ CHANGELOG.markdown
@@ -2,6 +2,8 @@
 ----
 * Fixed the `Hashable` instance for `Arg` to only hash the first argument,
   lest equal values have different hashes.
+* Add `Lift (NonEmpty a)` instance.
+* `Data.List.NonEmpty` is unconditionally *Trustworthy*
 
 0.18.5 [2018.07.02]
 -------------------